Do some modification based on review comments
